# Conflicts:
#	src/adventure/Booking.java

+++ src/adventure/Activity.java
@@ -9,7 +9,7 @@ public class Activity {
     //Variabler for aktiviteter
 
     private String activityName;
-    private int price;
+    private double price;
     private int requiredAge;
     private String description;
 
@@ -25,7 +25,7 @@ public class Activity {
     public Activity() {
     }
 
-    public Activity(String activityName, int price, int requiredAge, String description, boolean bookedOrNot) {
+    public Activity(String activityName, double price, int requiredAge, String description, boolean bookedOrNot) {
         this.activityName = activityName;
         this.price = price;
         this.requiredAge = requiredAge;
@@ -43,11 +43,11 @@ public class Activity {
         this.activityName = activityName;
     }
 
-    public int getPrice() {
+    public double getPrice() {
         return price;
     }
 
-    public void setPrice(int price) {
+    public void setPrice(double price) {
         this.price = price;
     }
 
@@ -89,7 +89,7 @@ public class Activity {
         activityName = sc.nextLine();
         System.out.println("You have created " + activityName + "\n");
         System.out.println("Choose a price:");
-        price = sc.nextInt();
+        price = sc.nextDouble();
         System.out.println("Choose a required age: ");
         requiredAge = sc.nextInt();
         System.out.println("Write a short description");
